# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/hash/alarms": {
            "get": {
                "description": "Get the hash of the alarms. Used by the iFrames to check updates and reload the iframe.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the hash of the alarms",
                "parameters": [
                    {
                        "type": "string",
                        "example": "netdata,radarr,sonarr",
                        "description": "Alarms to show. Available values: netdata, radarr, lidarr, sonarr, prowlarr, speedtest-tracker, pihole, kavita, kaizoku, changedetectionio, backrest",
                        "name": "alarms",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "example": false,
                        "description": "Sort alarms in descending order. Defaults to false.",
                        "name": "sort_desc",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": false,
                        "description": "Show only alarms that match or not the regex. Default to true.",
                        "name": "regex_include",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": false,
                        "description": "Show viewed alarms from changedetection.io. Defaults to true.",
                        "name": "changedetectionio_show_viewed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.hashResponse"
                        }
                    }
                }
            }
        },
        "/hash/cinemark": {
            "get": {
                "description": "Get the hash of the Cinemark movies. Used by the iFrames to check updates and reload the iframe.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the hash of the Cinemark movies",
                "parameters": [
                    {
                        "type": "string",
                        "example": "715, 1222, 4555",
                        "description": "The theater IDs to get movies from. It used to be easy to get, but now it's harder. To get it, you need to access the cinemark site, select a theater, open your browser developer console, go to the ",
                        "name": "theaterIds",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 5,
                        "description": "Limits the number of items in the iFrame.",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.hashResponse"
                        }
                    }
                }
            }
        },
        "/hash/linkwarden": {
            "get": {
                "description": "Get the hash of the Linkwarden bookmarks. Used by the iFrames to check updates and reload the iframe.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the hash of the Linkwarden bookmarks",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Get bookmarks only from this collection. You can get the collection ID by going to the collection page. The ID should be on the URL. The ID of the default collection **Unorganized** is 1 because the URL is https://domain.com/collections/1.",
                        "name": "collectionId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 5,
                        "description": "Limits the number of items in the iFrame.",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.hashResponse"
                        }
                    }
                }
            }
        },
        "/hash/media_releases": {
            "get": {
                "description": "Get the hash of the media releases. Used by the iFrames to check updates and reload the iframe.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the hash of media releases",
                "parameters": [
                    {
                        "type": "string",
                        "example": "physical",
                        "description": "Filter movies get from Radarr. Can be 'inCinemas', 'physical', or 'digital'. Defaults to 'inCinemas'",
                        "name": "radarrReleaseType",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": true,
                        "description": "Specify if show unmonitored media. Defaults to false.",
                        "name": "showUnmonitored",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.hashResponse"
                        }
                    }
                }
            }
        },
        "/hash/media_requests": {
            "get": {
                "description": "Get the hash of the media requests. Used by the iFrames to check updates and reload the iframe.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the hash of media requests",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 5,
                        "description": "Limits the number of items in the iFrame.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "all",
                        "description": "Filters for request status and media status. Available values: all, approved, available, pending, processing, unavailable, failed, deleted, completed, allavaliable (showMedia=true). Defaults to all",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "added",
                        "description": "Available values: added, modified, mediaAdded (showMedia=true). Defaults to added",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1",
                        "description": "If specified, only requests from that particular overseerr user ID will be returned.",
                        "name": "requestedByOverseerr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "1",
                        "description": "If specified, only requests from that particular jellyseerr user ID will be returned.",
                        "name": "requestedByJellyseerr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "true",
                        "description": "If true, shows the requests' media data, not the requests and media data. Defaults to false.",
                        "name": "showMedia",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.hashResponse"
                        }
                    }
                }
            }
        },
        "/hash/uptimekuma": {
            "get": {
                "description": "Get the hash of the Uptime Kuma sites status. Used by the iFrames to check updates and reload the iframe.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the hash of the Uptime Kuma sites status",
                "parameters": [
                    {
                        "type": "string",
                        "example": "uptime-kuma-slug",
                        "description": "You need to create a status page in Uptime Kuma and select which sites/services this status page will show. While creating the status page, it'll request **you** to create a slug, after creating the status page, provide this slug here. This iFrame will show data only of the sites/services of this specific status page!",
                        "name": "slug",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.hashResponse"
                        }
                    }
                }
            }
        },
        "/hash/vikunja": {
            "get": {
                "description": "Get the hash of the Vikunja tasks. Used by the iFrames to check updates and reload the iframe.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the hash of the Vikunja tasks",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 5,
                        "description": "Limits the number of items in the iFrame.",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Project ID to get tasks from. You can get it by going to the project page in Vikunja, the project ID should be on the URL. Example project page URL: https://vikunja.com/projects/2, the project ID is 2. Inbox tasks = 1, Favorite tasks = -1.",
                        "name": "project_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.hashResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "routes.hashResponse": {
            "type": "object",
            "properties": {
                "hash": {
                    "type": "string"
                }
            }
        },
...
    }
}